TemperatureData change to double

+++ src/examples/08_module/temperature.h
@@ -11,7 +11,7 @@ class Temperature
 public:
     Temperature(int h, double r) : hour(h), reading(r){/*empty code block*/}
     int get_hour()const{return hour;}
-    int get_reading()const{return reading; }
+    double get_reading()const{return reading; }
 
 private:
     int hour;
